refactored all components into subfolders
